Parallelize ebay crawling
Simply awaiting all futures sending get requests results in many
timeouts. Instead we limit the number of parallel requests to the number
of threds the CPU can handle.

This is generally more reliable but still encounters some timeout errors
for the GET requests. To be certain that every result has been gotten,
we increase the number of retries if a failure occures. We shouldn't run
into timeouts that often but it doesn't hurt to be on the safe side with
this.