Search each set number only once
A portfolio can contain a set number many times. That means searching
for each set number will look for some set numbers multiple times.

Instead we could search each set only once, store the result in a map,
and iterate over the original list (with duplicates) and look up the
price in the map.

For the current portfolio this cuts the runtime in half.