Improve error propagation
Report error for rows which can't be parsed correctly.
Report error when crawling produces no usable results.
Instead of having one generic place holder which will be used for every
row which encountered an error at one point in time we will print the
specific error for each row.

Also rows which have neither price nor set number will result in an
empty row in the result csv. We are not allowed to drop ant rows.

Some rows in the portfolios report some set number or price cells to be
popolated which don't show up in libre office calc. This is a known
issue currentlu without any workaround.